requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
anykeystore==0.2
apex==0.9.10.dev0
args==0.1.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==15.1.0
beautifulsoup4==4.13.5
bitsandbytes==0.47.0
blobfile==3.1.0
braceexpand==0.1.7
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
cachetools==5.5.2
certifi==2026.1.4
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.3.1
clint==0.5.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.2
cryptacular==1.6.2
cycler==0.12.1
datasets==4.1.1
decorator==5.2.1
decord==0.6.0
deepspeed==0.17.6
defusedxml==0.7.1
diffusers==0.35.1
dill==0.4.0
distro==1.9.0
dm-tree==0.1.9
easydict==1.13
einops==0.8.1
einops-exts==0.0.4
et_xmlfile @ file:///croot/et_xmlfile_1756462486635/work
exceptiongroup==1.3.1
executing==2.2.1
fairscale==0.4.13
filelock==3.20.3
fire==0.4.0
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.5.8/flash_attn-2.5.8+cu122torch2.3cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=13454dd3d37cf173649bd389b84614b8072fb283f8f2fd23a65ab66caafc304b
flatbuffers==25.12.19
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2026.1.0
ftfy==6.3.1
func_timeout==4.3.5
fvcore==0.1.5.post20221221
gast==0.7.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-auth==2.40.3
google-genai==1.2.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
hjson==3.1.0
hpsv2==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hupper==1.12.1
icecream==2.1.8
idna==3.11
imageio==2.31.1
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.1.0
iopath==0.1.10
jax==0.6.2
jaxlib==0.6.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
jiter==0.11.0
joblib==1.5.2
keras==3.12.1
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llava @ git+https://github.com/LLaVA-VL/LLaVA-NeXT.git@e9835311c6f515a13702eb7a7750fcd936f65ed8
llvmlite==0.45.0
lxml==6.0.1
Markdown==3.10.1
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml_dtypes==0.5.4
moviepy==2.2.1
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
namex==0.1.0
networkx @ file:///croot/networkx_1737039604450/work
ninja==1.13.0
nltk==3.9.1
numba==0.62.0
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
open_clip_torch==3.1.0
openai==1.108.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl @ file:///croot/openpyxl_1736366913861/work
opt_einsum==3.4.0
optree==0.18.0
oryx==0.2.9
packaging==26.0
pandas==2.3.2
parameterized==0.9.0
PasteDeploy==3.1.0
pbkdf2==1.3
peft==0.5.0
Pillow==10.0.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
portalocker==3.2.0
proglog==0.1.12
propcache==0.3.2
protobuf==6.33.5
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser @ file:///home/task_175749556993857/conda-bld/pycparser_1757496041120/work
pycryptodomex==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.4
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytorchvideo @ git+https://github.com/linzhiqiu/pytorchvideo.git@6cdc929315aab1b5674b6dcf73b16ec99147735f
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.11
regex==2026.1.15
repoze.sendmail==4.4.1
requests @ file:///croot/requests_1756709366904/work
requests-oauthlib==2.0.0
rich==14.3.1
rouge_score==0.1.2
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.38.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.8
soxr==1.0.0
SQLAlchemy==2.0.43
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-probability==0.25.0
termcolor==2.4.0
threadpoolctl==3.6.0
tiktoken==0.11.0
timm==1.0.19
tokenizers==0.21.4
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
TorchCodec==0.1.0
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transaction==5.0
transformers==4.51.3
translationstring==1.4
triton==3.1.0
typer-slim==0.21.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1750775463400/work
uvicorn==0.35.0
velruse==1.1.1
venusian==3.1.1
wandb==0.22.0
wcwidth==0.2.13
webdataset==1.0.2
WebOb==1.8.9
websockets==14.2
Werkzeug==3.1.5
wrapt==2.0.1
WTForms==3.2.1
wtforms-recaptcha==0.3.2
xformers==0.0.29.post1
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.1
zipp==3.23.0
zope.deprecation==6.0
zope.interface==8.0
zope.sqlalchemy==4.0
